image section changes

+++ app/Http/Controllers/ImageSectionController.php
@@ -55,7 +55,9 @@ class ImageSectionController extends Controller
             // Storage::disk('public-storage')->delete($section->section_image);
             // $data['section_image'] = $request->file('section_image')->store('uploads/section_images', 'public-storage');
             // Delete the old image from storage
-            Storage::disk('public-storage')->delete($section->section_image);
+            if($section->section_image !== "/uploads/default-images/blank-image-2.jpg" && $section->section_image !== "/uploads/default-images/blank-image-1.jpg") {
+                Storage::disk('public-storage')->delete($section->section_image);
+            }
             
             // Store the new image and get the relative path
             $relativePath = $request->file('section_image')->store('uploads/section_images', 'public-storage');
@@ -83,7 +85,9 @@ class ImageSectionController extends Controller
         $section = ImageSection::findOrFail($sectionId);
         
         // Delete the image associated with the section
-        Storage::disk('public-storage')->delete($section->section_image);
+        if($section->section_image !== "/uploads/default-images/blank-image-2.jpg" && $section->section_image !== "/uploads/default-images/blank-image-1.jpg") {
+            Storage::disk('public-storage')->delete($section->section_image);
+        }
         
         $section->delete();
